+++ front_end/src/pages/Accounts/AccountsAddPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react/addons';
-import { pageVisit } from 'actions/AppActions';
+import { pageVisit } from 'actions/App/AppActions';
 import { connect } from 'react-redux';
 
 require('styles/Accounts/accounts.scss');
