+++ front_end/src/pages/Accounts/AccountsPage.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react/addons';
 import { connect } from 'react-redux';
 
-import { pageVisit } from 'actions/AppActions';
+import { pageVisit } from 'actions/App/AppActions';
 import { saveAccount, fetchAccounts } from 'actions/Accounts/AccountActions';
 import AccountList from 'components/Accounts/AccountList/AccountList';
 import AccountForm from 'components/Accounts/AccountAdd/AccountForm';
